refactor: 让插件和工具模块完整拥有各自 Tauri 边界
