+++ desktop/src-tauri/src/main.rs
@@ -5279,38 +5279,6 @@ async fn settings_provider_model_save(
 }
 
 #[tauri::command]
-async fn settings_tools_get(
-    window: WebviewWindow,
-    shell: State<'_, product_shell::ProductShellState>,
-    lifecycle: State<'_, ShellLifecycleState>,
-) -> Result<Value, String> {
-    product_shell::validate_settings_window(&window)?;
-    let handle = settings_core_handle(&lifecycle)?;
-    let window_generation = shell.generation()?;
-    let core_generation_id = handle
-        .available_generation_id()
-        .map_err(str::to_string)?
-        .ok_or_else(|| "SETTINGS_CORE_UNAVAILABLE".to_string())?;
-    let response = dispatch_settings_request(
-        handle.clone(),
-        None,
-        "tools.settings.get",
-        json!({}),
-        std::time::Duration::from_secs(3),
-    )
-    .await?;
-    assert_settings_identity(&shell, &handle, window_generation, &core_generation_id)?;
-    let mut payload = settings_response_payload(response)?;
-    tool_settings::validate_snapshot(&payload, false)?;
-    let object = payload
-        .as_object_mut()
-        .ok_or_else(|| "TOOLS_SETTINGS_RESPONSE_INVALID".to_string())?;
-    object.insert("windowGeneration".to_string(), json!(window_generation));
-    object.insert("coreGenerationId".to_string(), json!(core_generation_id));
-    Ok(payload)
-}
-
-#[tauri::command]
 async fn settings_screen_awareness_get(
     window: WebviewWindow,
     shell: State<'_, product_shell::ProductShellState>,
@@ -5379,317 +5347,6 @@ async fn settings_screen_awareness_save(
         );
     }
     Ok(payload)
-}
-
-#[tauri::command]
-async fn settings_tools_save(
-    window: WebviewWindow,
-    window_generation: u64,
-    core_generation_id: String,
-    settings: Value,
-    shell: State<'_, product_shell::ProductShellState>,
-    lifecycle: State<'_, ShellLifecycleState>,
-) -> Result<Value, String> {
-    product_shell::validate_settings_window(&window)?;
-    tool_settings::validate_draft(&settings)?;
-    let handle = settings_core_handle(&lifecycle)?;
-    assert_settings_identity(&shell, &handle, window_generation, &core_generation_id)?;
-    let response = dispatch_settings_request(
-        handle.clone(),
-        None,
-        "tools.settings.save",
-        json!({"settings": settings}),
-        std::time::Duration::from_secs(5),
-    )
-    .await?;
-    let payload = settings_response_payload(response)?;
-    assert_settings_identity(&shell, &handle, window_generation, &core_generation_id)?;
-    tool_settings::validate_snapshot(&payload, true)?;
-    Ok(payload)
-}
-
-#[tauri::command]
-async fn settings_plugins_get(
-    window: WebviewWindow,
-    shell: State<'_, product_shell::ProductShellState>,
-    lifecycle: State<'_, ShellLifecycleState>,
-) -> Result<Value, String> {
-    product_shell::validate_settings_window(&window)?;
-    let handle = settings_core_handle(&lifecycle)?;
-    let window_generation = shell.generation()?;
-    let core_generation_id = handle
-        .available_generation_id()
-        .map_err(str::to_string)?
-        .ok_or_else(|| "SETTINGS_CORE_UNAVAILABLE".to_string())?;
-    let response = dispatch_settings_request(
-        handle.clone(),
-        None,
-        "plugins.settings.get",
-        json!({}),
-        std::time::Duration::from_secs(4),
-    )
-    .await?;
-    assert_settings_identity(&shell, &handle, window_generation, &core_generation_id)?;
-    let mut payload = settings_response_payload(response)?;
-    plugin_settings::validate_snapshot(&payload, false)?;
-    let object = payload
-        .as_object_mut()
-        .ok_or_else(|| "PLUGIN_SETTINGS_RESPONSE_INVALID".to_string())?;
-    object.insert("windowGeneration".to_string(), json!(window_generation));
-    object.insert("coreGenerationId".to_string(), json!(core_generation_id));
-    Ok(payload)
-}
-
-#[tauri::command]
-async fn settings_plugins_save(
-    window: WebviewWindow,
-    window_generation: u64,
-    core_generation_id: String,
-    plugin_id: String,
-    section_id: String,
-    values: Value,
-    shell: State<'_, product_shell::ProductShellState>,
-    lifecycle: State<'_, ShellLifecycleState>,
-) -> Result<Value, String> {
-    product_shell::validate_settings_window(&window)?;
-    plugin_settings::validate_settings_save_request(&plugin_id, &section_id, &values)?;
-    let handle = settings_core_handle(&lifecycle)?;
-    assert_settings_identity(&shell, &handle, window_generation, &core_generation_id)?;
-    let response = dispatch_settings_request(
-        handle.clone(),
-        None,
-        "plugins.settings.save",
-        json!({"pluginId": plugin_id, "sectionId": section_id, "values": values}),
-        std::time::Duration::from_secs(8),
-    )
-    .await?;
-    let payload = settings_response_payload(response)?;
-    assert_settings_identity(&shell, &handle, window_generation, &core_generation_id)?;
-    plugin_settings::validate_settings_save_result(&payload)?;
-    Ok(payload)
-}
-
-#[tauri::command]
-async fn settings_plugins_enabled_set(
-    window: WebviewWindow,
-    window_generation: u64,
-    core_generation_id: String,
-    revision: String,
-    install_id: String,
-    enabled: bool,
-    shell: State<'_, product_shell::ProductShellState>,
-    lifecycle: State<'_, ShellLifecycleState>,
-) -> Result<Value, String> {
-    product_shell::validate_settings_window(&window)?;
-    plugin_settings::validate_enabled_request(&revision, &install_id)?;
-    let handle = settings_core_handle(&lifecycle)?;
-    assert_settings_identity(&shell, &handle, window_generation, &core_generation_id)?;
-    let response = dispatch_settings_request(
-        handle.clone(),
-        None,
-        "plugins.enabled.set",
-        json!({"revision": revision, "installId": install_id, "enabled": enabled}),
-        std::time::Duration::from_secs(12),
-    )
-    .await?;
-    let mut payload = settings_response_payload(response)?;
-    assert_settings_identity(&shell, &handle, window_generation, &core_generation_id)?;
-    plugin_settings::validate_management_result(&payload)?;
-    if payload.get("managementAction").and_then(Value::as_str) != Some("enabled_changed")
-        || payload.get("installId").and_then(Value::as_str) != Some(install_id.as_str())
-    {
-        return Err("PLUGIN_MANAGEMENT_RESPONSE_INVALID".to_string());
-    }
-    let object = payload
-        .as_object_mut()
-        .ok_or_else(|| "PLUGIN_MANAGEMENT_RESPONSE_INVALID".to_string())?;
-    object.insert("windowGeneration".to_string(), json!(window_generation));
-    object.insert("coreGenerationId".to_string(), json!(core_generation_id));
-    Ok(payload)
-}
-
-#[tauri::command]
-async fn settings_plugins_action(
-    window: WebviewWindow,
-    window_generation: u64,
-    core_generation_id: String,
-    plugin_id: String,
-    section_id: String,
-    action_id: String,
-    values: Value,
-    shell: State<'_, product_shell::ProductShellState>,
-    lifecycle: State<'_, ShellLifecycleState>,
-) -> Result<Value, String> {
-    product_shell::validate_settings_window(&window)?;
-    let handle = settings_core_handle(&lifecycle)?;
-    assert_settings_identity(&shell, &handle, window_generation, &core_generation_id)?;
-    let response = dispatch_settings_request(
-        handle.clone(),
-        None,
-        "plugins.settings.action",
-        json!({"pluginId": plugin_id, "sectionId": section_id, "actionId": action_id, "values": values}),
-        std::time::Duration::from_secs(5),
-    )
-    .await?;
-    let payload = settings_response_payload(response)?;
-    assert_settings_identity(&shell, &handle, window_generation, &core_generation_id)?;
-    plugin_settings::validate_action_result(&payload)?;
-    Ok(payload)
-}
-
-#[tauri::command]
-async fn settings_plugins_install(
-    window: WebviewWindow,
-    window_generation: u64,
-    core_generation_id: String,
-    revision: String,
-    source_kind: String,
-    shell: State<'_, product_shell::ProductShellState>,
-    lifecycle: State<'_, ShellLifecycleState>,
-) -> Result<Value, String> {
-    product_shell::validate_settings_window(&window)?;
-    let handle = settings_core_handle(&lifecycle)?;
-    assert_settings_identity(&shell, &handle, window_generation, &core_generation_id)?;
-    let selected = match source_kind.as_str() {
-        "zip" => {
-            rfd::AsyncFileDialog::new()
-                .add_filter("Sakura 插件 ZIP", &["zip"])
-                .pick_file()
-                .await
-        }
-        "folder" => rfd::AsyncFileDialog::new().pick_folder().await,
-        _ => return Err("PLUGIN_INSTALL_SOURCE_INVALID".to_string()),
-    };
-    let Some(selected) = selected else {
-        return Ok(json!({"cancelled": true}));
-    };
-    assert_settings_identity(&shell, &handle, window_generation, &core_generation_id)?;
-    let source_path = selected
-        .path()
-        .to_str()
-        .filter(|value| !value.is_empty() && value.len() <= 4096)
-        .ok_or_else(|| "PLUGIN_INSTALL_SOURCE_INVALID".to_string())?;
-    if source_kind == "zip"
-        && selected
-            .path()
-            .extension()
-            .and_then(|value| value.to_str())
-            .is_none_or(|value| !value.eq_ignore_ascii_case("zip"))
-    {
-        return Err("PLUGIN_INSTALL_SOURCE_INVALID".to_string());
-    }
-    let response = dispatch_settings_request(
-        handle.clone(),
-        None,
-        "plugins.install",
-        json!({
-            "revision": revision,
-            "sourceKind": source_kind,
-            "sourcePath": source_path,
-        }),
-        std::time::Duration::from_secs(30),
-    )
-    .await?;
-    let mut payload = settings_response_payload(response)?;
-    assert_settings_identity(&shell, &handle, window_generation, &core_generation_id)?;
-    plugin_settings::validate_management_result(&payload)?;
-    if payload.get("managementAction").and_then(Value::as_str) != Some("installed") {
-        return Err("PLUGIN_MANAGEMENT_RESPONSE_INVALID".to_string());
-    }
-    let object = payload
-        .as_object_mut()
-        .ok_or_else(|| "PLUGIN_MANAGEMENT_RESPONSE_INVALID".to_string())?;
-    object.insert("windowGeneration".to_string(), json!(window_generation));
-    object.insert("coreGenerationId".to_string(), json!(core_generation_id));
-    Ok(payload)
-}
-
-#[tauri::command]
-async fn settings_plugins_uninstall(
-    window: WebviewWindow,
-    window_generation: u64,
-    core_generation_id: String,
-    revision: String,
-    install_id: String,
-    shell: State<'_, product_shell::ProductShellState>,
-    lifecycle: State<'_, ShellLifecycleState>,
-) -> Result<Value, String> {
-    product_shell::validate_settings_window(&window)?;
-    let handle = settings_core_handle(&lifecycle)?;
-    assert_settings_identity(&shell, &handle, window_generation, &core_generation_id)?;
-    let response = dispatch_settings_request(
-        handle.clone(),
-        None,
-        "plugins.uninstall",
-        json!({"revision": revision, "installId": install_id}),
-        std::time::Duration::from_secs(30),
-    )
-    .await?;
-    let mut payload = settings_response_payload(response)?;
-    assert_settings_identity(&shell, &handle, window_generation, &core_generation_id)?;
-    plugin_settings::validate_management_result(&payload)?;
-    if payload.get("managementAction").and_then(Value::as_str) != Some("uninstalled")
-        || payload.get("installId").and_then(Value::as_str) != Some(install_id.as_str())
-    {
-        return Err("PLUGIN_MANAGEMENT_RESPONSE_INVALID".to_string());
-    }
-    let object = payload
-        .as_object_mut()
-        .ok_or_else(|| "PLUGIN_MANAGEMENT_RESPONSE_INVALID".to_string())?;
-    object.insert("windowGeneration".to_string(), json!(window_generation));
-    object.insert("coreGenerationId".to_string(), json!(core_generation_id));
-    Ok(payload)
-}
-
-#[tauri::command]
-async fn settings_plugins_collection(
-    window: WebviewWindow,
-    window_generation: u64,
-    core_generation_id: String,
-    operation: String,
-    plugin_id: String,
-    section_id: String,
-    collection_id: String,
-    payload: Value,
-    shell: State<'_, product_shell::ProductShellState>,
-    lifecycle: State<'_, ShellLifecycleState>,
-) -> Result<Value, String> {
-    product_shell::validate_settings_window(&window)?;
-    plugin_settings::validate_collection_request(
-        &operation,
-        &plugin_id,
-        &section_id,
-        &collection_id,
-        &payload,
-    )?;
-    let handle = settings_core_handle(&lifecycle)?;
-    assert_settings_identity(&shell, &handle, window_generation, &core_generation_id)?;
-    let mut request_payload = payload
-        .as_object()
-        .cloned()
-        .ok_or_else(|| "PLUGIN_COLLECTION_REQUEST_INVALID".to_string())?;
-    request_payload.insert("pluginId".to_string(), json!(plugin_id));
-    request_payload.insert("sectionId".to_string(), json!(section_id));
-    request_payload.insert("collectionId".to_string(), json!(collection_id));
-    let request_name = match operation.as_str() {
-        "query" => "plugins.collection.query",
-        "create" => "plugins.collection.create",
-        "update" => "plugins.collection.update",
-        "delete" => "plugins.collection.delete",
-        _ => return Err("PLUGIN_COLLECTION_REQUEST_INVALID".to_string()),
-    };
-    let response = dispatch_settings_request(
-        handle.clone(),
-        None,
-        request_name,
-        Value::Object(request_payload),
-        std::time::Duration::from_secs(5),
-    )
-    .await?;
-    let result = settings_response_payload(response)?;
-    assert_settings_identity(&shell, &handle, window_generation, &core_generation_id)?;
-    plugin_settings::validate_collection_result(&operation, &result)?;
-    Ok(result)
 }
 
 #[tauri::command]
@@ -8470,17 +8127,17 @@ fn main() {
             settings_provider_model_save,
             settings_provider_model_probe,
             settings_provider_model_cancel,
-            settings_tools_get,
-            settings_tools_save,
+            tool_settings::settings_tools_get,
+            tool_settings::settings_tools_save,
             settings_screen_awareness_get,
             settings_screen_awareness_save,
-            settings_plugins_get,
-            settings_plugins_save,
-            settings_plugins_enabled_set,
-            settings_plugins_action,
-            settings_plugins_install,
-            settings_plugins_uninstall,
-            settings_plugins_collection,
+            plugin_settings::settings_plugins_get,
+            plugin_settings::settings_plugins_save,
+            plugin_settings::settings_plugins_enabled_set,
+            plugin_settings::settings_plugins_action,
+            plugin_settings::settings_plugins_install,
+            plugin_settings::settings_plugins_uninstall,
+            plugin_settings::settings_plugins_collection,
             product_shell::resolve_settings_close,
             resolve_settings_exit
         ])
